tools/banking/requirements.txt
# Mind Protocol Banking Integration Dependencies

# JWT token generation
PyJWT==2.8.0
cryptography==41.0.7

# API client
requests==2.31.0

# Environment configuration
python-dotenv==1.0.0
